fix: support decimal ratings and improve profile/map features
프로필 및 평점 관련 개선사항:

Backend Changes:
- 평점 타입을 uint8에서 float32로 변경하여 소수점 지원 (예: 3.5점)
- migration_rating_to_float.sql 추가하여 DB 스키마 업데이트
- createMemoHandler, updateMemoHandler에서 float32 파싱 지원

Frontend Changes:
- ProfileViewModel: bio 변경 체크 로직 추가하여 자기소개 수정 가능하도록 수정
- ProfileScreen: 자기소개 필드 표시 추가
- EditMemoScreen: 메모 내용 입력 필드 추가
- MapViewModel: GPS 위치 없을 때 서울 중심 좌표를 기본값으로 사용하여 검색 개선

Database Migration:
- ALTER TABLE memos MODIFY COLUMN rating FLOAT 실행 필요

+++ backend/src/features/memo/model/request/createMemo.go
@@ -11,7 +11,7 @@ type ReqCreateMemo struct {
 	ImageURL        string                `json:"image_url"`
 	ImageFile       multipart.File        `json:"-"` // S3 업로드용 파일
 	ImageHeader     *multipart.FileHeader `json:"-"` // 파일 메타데이터
-	Rating          uint8                 `json:"rating"`
+	Rating          float32               `json:"rating"` // Changed from uint8 to float32 for decimal support
 	IsPinned        bool                  `json:"is_pinned"`
 	Latitude        *float64              `json:"latitude"`
 	Longitude       *float64              `json:"longitude"`
